replace OldEGLRuntime.cpp
