+++ src/Entry/EGL/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(egl_runtime NewEGLRuntime.cpp)
+add_library(egl_runtime EGLRuntime.cpp)
 set_target_properties(egl_runtime PROPERTIES POSITION_INDEPENDENT_CODE ON)
 target_link_libraries(egl_runtime OpenGL::EGL vgg_layer)
